abstract details panel to add admin side
