+++ frontend/src/components/AdminDetailsPanel/index.js
@@ -1,0 +1,9 @@
+function AdminDetailsPanel() {
+    return (
+        <div>
+            Schedules
+        </div>
+    )
+}
+
+export default AdminDetailsPanel;
