+++ frontend/src/components/VolunteerDetailsPage/index.css
@@ -1,0 +1,208 @@
+@import '../../styles.css';  /*Imports custom fonts*/
+
+.panel-button-icon {
+     padding: 0px;
+     background-color: transparent;
+     border: none;
+     cursor: pointer;
+}
+
+.panel-button-icon:hover {
+     filter:brightness(0.6);
+}
+
+.panel-details {
+     margin-right: 36px; 
+     margin-left: 20px;
+     padding: 32px 0px;
+}
+
+.panel-titles {
+     min-width: 120px;
+     font-family: 'Roboto';
+     font-weight: 400;
+     font-size: 16px;
+     line-height: 21.6px;
+     color: #808080;
+}
+
+.my-shifts {
+     padding: 4px 12px;
+     width: fit-content;
+     border-radius: 4px;
+     background: rgba(24, 149, 49, 0.1);
+     color: var(--green);
+     font-weight: bold;
+     text-align: center;
+ }
+ 
+ .my-coverage-requests {
+     padding: 4px 12px;
+     width: fit-content;
+     border-radius: 4px;
+     background: rgba(203, 131, 36, 0.1);
+     color: var(--yellow);
+     font-weight: bold;
+     text-align: center;
+ }
+ 
+ .coverage {
+     padding: 4px 12px;
+     width: fit-content;
+     border-radius: 4px;
+     background: rgba(149, 32, 24, 0.1);
+     color: var(--error-red);
+     font-weight: bold;
+     text-align: center;
+ }
+
+.classTaken {
+     padding: 4px 12px;
+     width: fit-content;
+     border-radius: 4px;
+     background: rgb(211, 211, 211);
+     color: rgb(111, 111, 111);
+     font-weight: bold;
+     text-align: center;
+}
+
+.volunteersNeeded {
+     padding: 4px 12px;
+     width: fit-content;
+     border-radius: 4px;
+     background: rgba(89, 165, 213, 0.1);
+     color: var(--primary-blue);
+     font-weight: bold;
+     text-align: center;
+}
+
+.panel-details-shift {
+     display: flex;
+     flex-direction: column;
+     gap: 12px;
+}
+
+.panel-details-shift-row {
+     display: flex;
+     flex-direction: row;
+     align-items: center;
+}
+
+.panel-details-shift-right {
+     width: 264px;
+     font-family: 'Roboto';
+     font-weight: 400;
+     font-size: 16px;
+     display: flex; 
+     align-items: center;
+}
+
+.volunteer-profile {
+     width: 28px;
+     height: 28px;
+     background-color: #808080;
+     border-radius: 4px;
+     opacity: 0.3;
+     margin-right: 6px;
+}
+
+.panel-details-description {
+     margin-top: 20px;
+}
+
+.panel-description {
+     margin-top: 10px;
+     font-family: 'Roboto';
+     font-weight: 400;
+     font-size: 16px;
+     line-height: 21.6px;
+}
+
+.button-icons {
+     display: flex;
+     flex-direction: row;
+     gap: 4px;
+     width: fit-content;
+     margin-left: auto;
+     margin-right: 36px;
+     margin-top: 20px;
+}
+
+ /* Shift Details Panel */
+/* Dynamic Button Styling based on Shift Type */
+.panel-buttons {
+     padding-top: 2vh;
+     height: 4vh;
+     display: flex;
+     flex-direction: row;
+     gap: 12px;
+}
+
+.dynamic-button {
+     position: relative;
+     height: 100%;
+     border-radius: 8px;
+     border: 1px solid var(--primary-blue);
+     background: #FFF;
+     box-shadow: 0px 4px 16px 0px rgba(15, 17, 17, 0.05);
+     padding: 0 1rem;
+     display: flex;
+     align-items: center;
+ }
+ 
+ .dynamic-button:not(:disabled):hover {
+     cursor: pointer;
+     background: var(--button-blue);
+     border: none;
+ }
+ 
+ .dynamic-button:disabled {
+     background-color: #d3d3d3;
+     border: 1px solid #d3d3d3;
+     box-shadow: none;
+ }
+ 
+ /* Icon within Button */
+ .card-button-icon {
+     width: 1rem;
+     height: 1rem;
+     margin-right: .5rem;
+ }
+
+ .zoom-icon {
+     height: 1.5rem;
+     width: 1.5rem;
+     margin-right: 0.5rem; 
+ }
+
+ .email-icon {
+     margin-left: 0.5rem;
+ }
+ 
+ .join-class-button {
+     display: flex; 
+     flex-direction: row; 
+     align-items: center; 
+     padding: 0.5rem 1rem;
+     margin-top: 1rem;
+     text-decoration: none;
+     border-radius: 8px;
+     border: 1px solid var(--primary-blue);
+     background: #FFF;
+     cursor: pointer;
+     transition: background-color 0.3s ease;
+ }
+ 
+ .join-class-button a {
+     text-decoration: none;
+     color: var(--dark);
+     display: flex; 
+     flex-direction: row;
+     align-items: center; 
+ }
+ 
+ .join-class-button:hover {
+     cursor: pointer;
+     background: var(--button-blue);
+     border: none;
+ }
